requirements.txt
PyMuPDF==1.25.5
faiss-cpu==1.10.0
pacmap==0.8.0
langchain-community==0.3.21
langchain-openai==0.3.12
duckduckgo-search
beautifulsoup4==4.13.3
sentence-transformers==3.4.1
accelerate==1.5.2
fastapi
uvicorn
pydantic
python-dotenv